feat: add links to the multisig TX for the VIP 341
