--- vips/vip-338/bscmainnet.ts
@@ -1,30 +1,0 @@
-import { ProposalType } from "../../src/types";
-import { makeProposal } from "../../src/utils";
-
-const FAST_TRACK_TIMELOCK = "0x555ba73dB1b006F3f2C7dB7126d6e4343aDBce02";
-
-const vip338 = () => {
-  const meta = {
-    version: "v2",
-    title: "VIP-338 Whitelist XVS Vault Treasury in Ethereum",
-    description: ``,
-    forDescription: "Execute this proposal",
-    againstDescription: "Do not execute this proposal",
-    abstainDescription: "Indifferent to execution",
-  };
-
-  return makeProposal(
-    [
-      {
-        target: FAST_TRACK_TIMELOCK,
-        signature: "",
-        params: [],
-        value: "1",
-      },
-    ],
-    meta,
-    ProposalType.FAST_TRACK,
-  );
-};
-
-export default vip338;
